refractored top sites and added dummy cards
