+++ src/templateComponents/CardRow.tsx
@@ -1,29 +1,23 @@
-import React, { ReactElement, useEffect, useState } from "react";
-import { JsxElement } from "typescript";
-import Card from "./Card";
-
+import React, { ReactElement, useEffect, useRef, useState } from "react";
 
 function CardRow(props:any){
-    const [topSiteLinks, setTopSiteLinks] = useState(new Array<ReactElement>());
+   
+    const ref = useRef<any>(null)
 
-    function OutputTopSites(topSites: chrome.topSites.MostVisitedURL[]) {
-        topSites.forEach((site: chrome.topSites.MostVisitedURL) => {
-            setTopSiteLinks(prev => [...prev, <Card>
-                <a href={site.url}>{site.title}</a>
-            </Card>])
-            console.log(topSiteLinks);
-        })
-    }
-
-    useEffect(() => {
-        if (chrome.topSites !== undefined) {
-            chrome.topSites.get(OutputTopSites);
+    const scroll = (scrollOffset:number) => {
+        if(ref && ref.current){
+            ref.current.scrollLeft += scrollOffset;
         }
-    }, [])
-    
+        
+      };
     return(
-        <div className="grid grid-rows-1 grid-flow-col px-20 py-4 gap-10 overflow-x-auto">
-            {topSiteLinks}
+        <div>
+            <div className="grid grid-rows-1 grid-flow-col px-20 py-4 gap-10 overflow-x-hidden" ref={ref}>
+                {props.children}
+            </div>
+            <button onClick={()=>scroll(-200)}>left</button>
+            <button onClick={()=>scroll(200)}>Right</button>
+
         </div>
     )
 }
